refactor: monorepo — extract core to packages/core (pnpm workspace)
Restructure for multi-surface (vscode now, android next) without splitting the
repo. The core and surfaces change together (~half of recent commits touched
both), so a monorepo with workspace packages fits better than separate repos.

- src/ → packages/core/src, scripts/ → packages/core/test (git rename, history
  kept; the test files' "../src" imports stay valid since test/ and src/ are now
  siblings under packages/core).
- packages/core is the "@iqlabs-official/agent-sdk" package (name kept) and owns
  the claude/codex SDK deps; exports map src subpaths so surfaces import cleanly.
- Root package.json slimmed to a workspace root; pnpm-workspace.yaml added
  (packages/*, surfaces/*).
- surfaces/vscode imports the core as "@iqlabs-official/agent-sdk" (was
  ../../../src) via a workspace:* dep. tsup noExternal inlines the core into the
  single extension bundle (the SDKs stay external, resolved + spawned at runtime).
  vscode tsconfig → moduleResolution Bundler, stale ../../src include dropped.
- Fix (surfaced by the stricter resolution): QuickPickItem reserves `kind`, so
  the cloud-picker stashes our storage kind under `k` to avoid the type clash.

Also adds plans/mobile-strategy.md (Android-only mobile research; iOS shelved
because Apple blocks fork at the kernel — full reasoning recorded).

Verified: core tsc + vscode tsc clean, vscode build (2.25 MB, core inlined),
extension loads, all 7 core tests pass (incl. cross-CLI resume).
