+++ package.json
@@ -1,27 +1,11 @@
 {
-  "name": "@iqlabs-official/agent-sdk",
+  "name": "agentnet-monorepo",
   "version": "0.0.1",
-  "type": "module",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build": "tsup",
-    "test:run": "tsx scripts/test-runtime.ts"
-  },
-  "peerDependencies": {
-    "@iqlabs-official/solana-sdk": "^0.1.27",
-    "@solana/web3.js": "^1.98.0"
-  },
-  "devDependencies": {
-    "@iqlabs-official/solana-sdk": "^0.1.27",
-    "@solana/web3.js": "^1.98.0",
-    "@types/node": "^20.0.0",
-    "tsup": "^8.0.0",
-    "tsx": "^4.19.2",
-    "tweetnacl": "^1.0.3",
-    "typescript": "^5.6.0"
-  },
-  "dependencies": {
-    "@anthropic-ai/claude-agent-sdk": "^0.3.170",
-    "@openai/codex-sdk": "^0.139.0"
+    "test:run": "pnpm --filter @iqlabs-official/agent-sdk test:run",
+    "build:core": "pnpm --filter @iqlabs-official/agent-sdk build",
+    "build:vscode": "pnpm --filter agentnet-vscode build"
   }
 }
