+++ packages/core/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@iqlabs-official/agent-sdk",
+  "version": "0.0.1",
+  "type": "module",
+  "private": true,
+  "main": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./*": "./src/*.ts",
+    "./*.js": "./src/*.ts"
+  },
+  "scripts": {
+    "build": "tsup",
+    "test:run": "tsx test/test-runtime.ts"
+  },
+  "peerDependencies": {
+    "@iqlabs-official/solana-sdk": "^0.1.27",
+    "@solana/web3.js": "^1.98.0"
+  },
+  "devDependencies": {
+    "@iqlabs-official/solana-sdk": "^0.1.27",
+    "@solana/web3.js": "^1.98.0",
+    "@types/node": "^20.0.0",
+    "tsup": "^8.0.0",
+    "tsx": "^4.19.2",
+    "tweetnacl": "^1.0.3",
+    "typescript": "^5.6.0"
+  },
+  "dependencies": {
+    "@anthropic-ai/claude-agent-sdk": "^0.3.170",
+    "@openai/codex-sdk": "^0.139.0"
+  }
+}
