+++ packages/core/src/core/paths.ts
