+++ packages/core/src/runtime/detect.ts
