+++ packages/core/src/runtime/inject/claude.ts
