+++ surfaces/vscode/package.json
@@ -36,6 +36,7 @@
     "typescript": "^5.6.0"
   },
   "dependencies": {
+    "@iqlabs-official/agent-sdk": "workspace:*",
     "dompurify": "^3.4.9",
     "marked": "^18.0.5",
     "tweetnacl": "^1.0.3"
